Delegate to :ssl for setopts/2 when transport is tls
Previously we always called `:inet.setopts/2` when a timeout was detected
to set `{linger, {true, 0}}` to immediately tear down the socket before
closing the connection. In doing so and in the case of tls the sslsocket
structure gets passed to `:prim_inet.setopts/2` where the first argument
is expected to be a port identifier.